feat: update styles for logs detail view
