+++ frontend/src/container/LogDetailedView/Overview.styles.scss
@@ -14,9 +14,10 @@
 	}
 
 	.log-switch {
+		height: 36px;
 		display: flex;
-		justify-content: space-between;
 		align-items: center;
+		border-top: 1px solid var(--bg-slate-500);
 
 		.wrap-word-switch {
 			display: flex;
@@ -48,8 +49,9 @@
 		border-bottom: 1px solid var(--bg-slate-500);
 
 		.ant-collapse-header {
+			align-items: center;
 			border-radius: 2px;
-			background: rgba(113, 144, 249, 0.08);
+			background: rgba(171, 189, 255, 0.04);
 
 			padding: 8px;
 		}
@@ -84,6 +86,24 @@
 
 	.ant-table-wrapper .ant-table-cell {
 		padding: 14px 14px;
+	}
+
+	.attribute-tab-header {
+		display: flex;
+		align-items: center;
+		justify-content: space-between;
+
+		.action-btn {
+			border-radius: 2px;
+			border: 1px solid var(--bg-slate-400);
+			display: flex;
+			align-items: center;
+			justify-content: center;
+
+			background: none;
+			border: none;
+			box-shadow: none;
+		}
 	}
 }
